requirements.txt
streamlit>=1.28.0
langchain-community>=0.0.29
langchain-core>=0.1.33
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.4
faiss-cpu>=1.7.4
gspread>=6.0.0
google-generativeai>=0.3.0
langchain-google-genai
pypdf>=3.17.4
sentence-transformers>=2.2.2
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.2
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
